requirements.txt
matplotlib==3.9.1
numpy==1.26.4
opencv-python==4.10.0.84
optuna==3.6.1
optuna-integration==3.6.0
plotly==5.23.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.4
wandb==0.17.6
